Move join pool tests into own file (#1781)
